refactor: move feature-specific use* hooks from entities to features layer
